+++ src/util/string_util.cpp
@@ -103,11 +103,11 @@ std::string StringUtil::Prefix(const std::string &str,
   return (os.str());
 }
 
-std::string StringUtil::FormatSize(long bytes) {
-  double BASE = 1024;
-  double KB = BASE;
-  double MB = KB * BASE;
-  double GB = MB * BASE;
+std::string StringUtil::FormatSize(size_t bytes) {
+  size_t BASE = 1024;
+  size_t KB = BASE;
+  size_t MB = KB * BASE;
+  size_t GB = MB * BASE;
 
   std::ostringstream os;
